add detail page

+++ src/components/dropdown/products.tsx
@@ -1,13 +1,23 @@
 'use client'
-
+import { useTransition } from 'react'
+import { useRouter } from 'next/navigation'
 import { DropdownMenuItem } from '@/components/ui/dropdown-menu'
 import { CheckIcon, TrashIcon, XIcon } from 'lucide-react'
-import { useRouter } from 'next/navigation'
-import { useTransition } from 'react'
+
 import {
   deleteProduct,
   toggleProductAvailability
 } from '@/lib/actions/admin/products'
+import {
+  AlertDialog,
+  AlertDialogTrigger,
+  AlertDialogContent,
+  AlertDialogTitle,
+  AlertDialogDescription,
+  AlertDialogCancel,
+  AlertDialogAction
+} from '@/components/ui/alert-dialog'
+import { AlertDialogHeader, AlertDialogFooter } from '../ui/alert-dialog'
 
 export function ActiveToggleDropdownItem({
   id,
@@ -53,18 +63,38 @@ export function DeleteDropdownItem({
   const [pending, startTransition] = useTransition()
   const router = useRouter()
   return (
-    <DropdownMenuItem
-      variant="destructive"
-      disabled={disabled || pending}
-      onClick={() => {
-        startTransition(async () => {
-          await deleteProduct(id)
-          router.refresh()
-        })
-      }}
-    >
-      <TrashIcon className="mr-2 size-4" />
-      <span>Delete</span>
-    </DropdownMenuItem>
+    <AlertDialog>
+      <AlertDialogTrigger asChild>
+        <DropdownMenuItem
+          variant="destructive"
+          disabled={disabled || pending}
+        >
+          <TrashIcon className="mr-2 size-4" />
+          <span>Delete</span>
+        </DropdownMenuItem>
+      </AlertDialogTrigger>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+          <AlertDialogDescription>
+            This action cannot be undone. This will permanently delete the order
+            and remove data from our servers.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>Cancel</AlertDialogCancel>
+          <AlertDialogAction
+            onClick={async () =>
+              startTransition(async () => {
+                await deleteProduct(id)
+                router.refresh()
+              })
+            }
+          >
+            Continue
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   )
 }
